+++ moor/test/database_test.dart
@@ -18,7 +18,7 @@ class _FakeDb extends GeneratedDatabase {
       },
       beforeOpen: (details) async {
         // this fake select query is verified via mocks
-        await customSelectQuery(
+        await customSelect(
                 'opened: ${details.versionBefore} to ${details.versionNow}')
             .get();
       },