+++ moor/test/utils/lazy_database_test.dart
@@ -52,7 +52,7 @@ void main() {
     final db = TodoDb(LazyDatabase(() => inner));
 
     // run a statement to make sure the database has been opened
-    await db.customSelectQuery('custom_select').get();
+    await db.customSelect('custom_select').get();
 
     verify(inner.databaseInfo = db);
   });