+++ moor/CHANGELOG.md
@@ -8,6 +8,8 @@
 - __Breaking__: Remove the second type variable on `Expression` and subclasses.
 - __Breaking__: Remove `InsertStatement.insertAll` (use batches instead) and `insert(orReplace: true)`
   (use `mode: InsertMode.orReplace` instead).
+- __Breaking__: Remove `customSelectStream` from `QueryEngine`. The `customSelect`
+  method now returns an `Selectable` (like `customSelectQuery`, which in turn has been deprecated).
 - Experimentally support IndexedDB to store sqlite data on the web
 
 ## 2.4.0